// frontend/src/core/api/generated/model/updatePlanRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */
import type { UpdatePlanRequestStatus } from './updatePlanRequestStatus'

export interface UpdatePlanRequest {
  /**
   * @minLength 0
   * @maxLength 60
   */
  name: string
  /**
   * @minLength 0
   * @maxLength 160
   * @nullable
   */
  tagline?: string | null
  /**
   * @minimum 0
   * @nullable
   */
  monthlyPrice?: number | null
  /**
   * @minimum 1
   * @nullable
   */
  maxUsers?: number | null
  /**
   * @minimum 1
   * @nullable
   */
  maxSites?: number | null
  highlighted: boolean
  visible: boolean
  status: UpdatePlanRequestStatus
  features: string[]
}